+++ src/test/e2e/commands/drivers.test.ts
@@ -15,7 +15,7 @@ suite("Drivers ...", function () {
    * Sets up the test suite before running any tests.
    */
   suiteSetup(async function () {
-    await LiquibaseGUITestUtils.openWorkspaceAndInitializeExtension();
+    await LiquibaseGUITestUtils.openWorkspace();
     //create a dummy jar file
     fs.writeFileSync(path.join(LiquibaseGUITestUtils.WORKSPACE_PATH, "dummy.jar"), "");
   });